исправления по заданию Модуляция
