+++ js/data/data.js
@@ -1,6 +1,6 @@
 import {getRandomPositiveInteger} from '../utils/get-random-positive-integer';
 
-const OBJESTS_TO_GEN = 25;
+const OBJECTS_TO_GEN = 25;
 const MESSAGES = [
   'Всё отлично!', 'В целом всё неплохо. Но не всё',
   'Когда вы делаете фотографию, хорошо бы убирать палец из кадра', 'В конце концов это просто непрофессионально.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
@@ -37,8 +37,18 @@ const createObject = (id) => {
   };
 };
 
-export {createComment, createObject,
-  OBJESTS_TO_GEN,
+const generateObjects = (count) => {
+  const objects = [];
+  for (let object = 1; object <= count; object++) {
+    objects.push(createObject(object));
+  }
+  return objects;
+};
+
+generateObjects(OBJECTS_TO_GEN);
+
+export {createComment, createObject, generateObjects,
+  OBJECTS_TO_GEN,
   MESSAGES,
   NAMES,
   LIKES_MIN,
